Change delete key IOCTL request structure to use ieee80211_keyix_t.
Ticket: http://madwifi.org/ticket/1702

WARNING: This changes userspace ABI - hostapd WILL need to be recompiled


git-svn-id: http://madwifi-project.org/svn/madwifi/trunk@3068 0192ed92-7a03-0410-a25b-9323aeb14dbd

+++ net80211/ieee80211_wireless.c
@@ -3400,7 +3400,7 @@ ieee80211_ioctl_delkey(struct net_device *dev, struct iw_request_info *info,
 
 	/* XXX: This cast can be removed when struct ieee80211req_del_key is 
 	 * fixed. */
-	if (dk->idk_keyix == (u_int8_t) IEEE80211_KEYIX_NONE) {
+	if (dk->idk_keyix == IEEE80211_KEYIX_NONE) {
 		struct ieee80211_node *ni;
 
 		ni = ieee80211_find_node(&ic->ic_sta, dk->idk_macaddr);
